Enable unconvert linter, fix errors (#1546)

+++ receiver/awsecscontainermetricsreceiver/awsecscontainermetrics/accumulator_test.go
@@ -23,7 +23,7 @@ import (
 
 func TestGetMetricsData(t *testing.T) {
 	v := uint64(1)
-	f := float64(1.0)
+	f := 1.0
 
 	memStats := make(map[string]uint64)
 	memStats["cache"] = v
